requirements_improved.txt
# ML Pipeline Requirements for Improved Disease Prediction System
# TensorFlow 2.12+ with Keras, OpenCV, and supporting libraries

# Core ML Framework
tensorflow>=2.12.0,<2.14.0
keras>=2.12.0
numpy>=1.22.0,<1.25.0

# Image Processing
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0

# Data Processing
pandas>=1.5.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Framework
Flask>=2.3.0
Flask-CORS>=4.0.0
Werkzeug>=2.3.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# GPU Support (optional, but recommended)
# Uncomment below if you have NVIDIA GPU
# tensorflow[and-cuda]>=2.12.0
# tensorflow-io-gcs-filesystem>=0.33.0

# Development Tools
jupyter>=1.0.0
ipython>=8.12.0

# Production Deployment
gunicorn>=21.2.0
python-dateutil>=2.8.2

# Monitoring & Logging
tqdm>=4.65.0

# Note: 
# - For CPU only: standard tensorflow package works fine
# - For GPU (NVIDIA): Ensure CUDA 11.x and cuDNN 8.x installed
# - For better performance: tensorflow-gpu>=2.12.0
# - Installation: pip install -r requirements_improved.txt
